Move submodules of internal to top level
