+++ src/matcher/unordered_vec_eq.rs
@@ -1,6 +1,6 @@
 use super::ArgumentMatcher;
-use crate::internal::argument::Argument;
-use crate::internal::fmt::MaybeDebugWrapper;
+use crate::argument::Argument;
+use crate::fmt::MaybeDebugWrapper;
 use nameof::name_of;
 use std::fmt::{self, Debug, Display};
